Rebecca: havent tested yet bc android studio is dying on me but started setting up navigation for creating editing and deleting courses
