+++ app/build.gradle
@@ -1,6 +1,6 @@
 plugins {
     id 'com.android.application'
-    id 'com.google.gms.google-services'
+    //id 'com.google.gms.google-services'
 }
 
 android {
@@ -33,7 +33,7 @@ android {
 }
 
 dependencies {
-
+/*
     classpath 'com.google.gms:google-services:4.3.13'
 
     // Import the Firebase BoM
@@ -43,6 +43,8 @@ dependencies {
     implementation 'com.google.firebase:firebase-analytics'
     // Add the dependencies for any other desired Firebase products
     // https://firebase.google.com/docs/android/setup#available-libraries
+
+ */
     implementation 'androidx.appcompat:appcompat:1.5.1'
     implementation 'com.google.android.material:material:1.7.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
