--- app/google-services.json
@@ -1,40 +1,0 @@
-{
-  "project_info": {
-    "project_number": "758080405649",
-    "firebase_url": "https://b07project-943e2-default-rtdb.firebaseio.com",
-    "project_id": "b07project-943e2",
-    "storage_bucket": "b07project-943e2.appspot.com"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:758080405649:android:fc7c04b7bbceb11e2cff30",
-        "android_client_info": {
-          "package_name": "com.android.application"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "758080405649-g73sq66o14lnugjrg92ouhbfe46mn4s3.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyCs7mpxJjc2FMdIFiYGNtaMt7VNyd8Zfhw"
-        }
-      ],
-      "services": {
-        "appinvite_service": {
-          "other_platform_oauth_client": [
-            {
-              "client_id": "758080405649-g73sq66o14lnugjrg92ouhbfe46mn4s3.apps.googleusercontent.com",
-              "client_type": 3
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
