+++ app/src/main/java/com/example/myapplication/AdminCourseManagement.java
@@ -1,0 +1,65 @@
+package com.example.myapplication;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.navigation.fragment.NavHostFragment;
+
+import com.example.myapplication.databinding.FragmentSecondBinding;
+
+public class AdminCourseManagement extends Fragment {
+
+    private FragmentSecondBinding binding;
+
+    @Override
+    public View onCreateView(
+            LayoutInflater inflater, ViewGroup container,
+            Bundle savedInstanceState
+    ) {
+
+        binding = FragmentSecondBinding.inflate(inflater, container, false);
+        return binding.getRoot();
+
+    }
+
+    public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+
+        binding.createcourse.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                NavHostFragment.findNavController(AdminCourseManagement.this)
+                        .navigate(R.id.action_SecondFragment_to_AdminCreate);
+            }
+        });
+
+        /* PLACEHOLDERS for when we add in SecondFragment_to_AdminEdit and SecondFragment_to_AdminDelete navigation
+        binding.editcourse.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                NavHostFragment.findNavController(AdminCourseManagement.this)
+                        .navigate(R.id.action_SecondFragment_to_AdminEdit);
+            }
+        });
+
+        binding.deletecourse.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                NavHostFragment.findNavController(AdminCourseManagement.this)
+                        .navigate(R.id.action_SecondFragment_to_AdminDelete);
+            }
+        });
+         */
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        binding = null;
+    }
+
+}
