+++ app/src/main/java/com/example/myapplication/AdminCreateCourse.java
@@ -1,0 +1,4 @@
+package com.example.myapplication;
+
+public class AdminCreateCourse {
+}
